task 0

+++ 0-bubble_sort.c
@@ -7,21 +7,23 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-    int i, j;
+	size_t i, j;
 
-    if (array == NULL || size < 2)
-		    return;
-    for (i = 0; i < size - 1; i++)
-    {
-        for (j = 0; j < size - 1 - i; j++)
-        {
-            if (array[j] > array[j + 1])
-            {
-                char temp = array[j];
-                array[j] = array[j + 1];
-                array[j + 1] = temp;
-                print_array(array, size);
-            }
-        }
-    }
+	if (array == NULL || size < 2)
+		return;
+
+	for (i = 0; i < size - 1; i++)
+	{
+		for (j = 0; j < size - 1 - i; j++)
+		{
+			if (array[j] > array[j + 1])
+			{
+				char temp = array[j];
+
+				array[j] = array[j + 1];
+				array[j + 1] = temp;
+				print_array(array, size);
+			}
+		}
+	}
 }
